Changes needed for OpenSSL 1.1
This change is the first cut at changes to OpenSC so it
can be compiled against OpenSSL 1.1 as well as previous versions.

OpenSSL 1.1 will be released later this year.

	https://www.openssl.org/policies/releasestrat.html

	https://www.openssl.org/news/changelog.html#x0

17 OpenSC routines need changes ranging from adding headers
for ossl_type.h, rsa.h and bn.h to rewriting code to use
changed interfaces.

5 routines that should be looked at closely are:

	src/libopensc/card-epass2003.c
	src/libopensc/pkcs15-itacns.c
	src/pkcs15init/pkcs15-oberthur-awp.c
	src/tools/pkcs15-init.c
	src/tools/sc-hsm-tool.c

src/tools/sc-hsm-tool.c needs so many changes that I did not
attempt to change it. If compiled against OpenSSL 1.1 it will
produce a dummy routine, and a warning message.
It uses the BIGNUM structure in line in many places, but BIGNUM
is now just a typedef and the actual structure is now hidden.
The sc-hsm developers need to make the changes.

I have compiles against OpenSSL 1.0.1g, 1.0.2d and 1.1.0-pre2-dev
from github.com Minimal testing has been done.

Next is engine and libp11  ECDH can be used from engine.