Handle C "inline" and "restrict"  keywords vor various compilers
simclist.h code is changed to either accept predefined inline
and restrict macros or set them based on the compiler being used.
for example, "configure" could define inline and/or restricted

The logic used is based on code in OpenSSL-1.1 e_os2.h

the one and only use of inline in card-dnie.c is removed as ther
is no need to have the function inline. This avaoids any inling issues.

If inline is needed in other OpeSC code, simclist_inline could be used.
simclist.h is included from internal.h

 Changes to be committed:
	modified:   ../common/simclist.c
	modified:   ../common/simclist.h
	modified:   card-dnie.c