asn1: parse 'securityCondition'
Currently the pkcs15-object data type, together with the existing level of parse of PKCS#15 xDF data,
assumes that protected object contains reference (AuthID) to protecting object in
CommonObjectAttributes.

The PKCS#15 standard and xDF data of existing PKCS#15 cards are more complicated then that:
AuthID of protecting object can be referenced per operation (and referenced in
'AuthID' identifier of related securityCondition),
it can also be present in logical operations 'and', 'or', 'not' combining the different
authentication methods (for ex. SM and/or PUK).

Full parse of such data is too complicated and needs considerable changes in pkcs15-object
data types.

Here is partial improvement:
pkcs15-object is protected if the reference to protecting object is found in:
- CommonObjectAttributes as 'AuthID' identifier (currently implemented);
- 'AuthID' identifier in one of SecurityCondtions (the first occurence);
- 'AuthID' from securityCondition operand of 'and' operation.