Add check_state functions
Firefox and other applications rely on C_GetSessionInfo to return if a login is needed.
They do this before doing other operations. OpenSC would return stale information because
it does not check with PCSC or the card to determine the state of the card.

This commit adds a stack of functions so C_GetSessionInfo gets the current state of the card.
PCSC only returns if the card has been reset or removed. A card driver can provide a
a check_state function that can issue commands to a card to get its true state.

    pkcs11-session.c: C_GetSessionInfo
    framework-pkcs15.c: pkcs15_check_state
    pkcs15.c: sc_pkcs15_check_state
    card.c: sc_check_state
    card-<driver>.c: <driver>_check_start

sc_pkcs15_check_state could also be used from minidriver or tokend if needed.

The <driver>_check_state function would be different for each card, and is optional.
Only a piv_check_state is included. It sets the AID back to the PIV AID then tests if
the cards logged state agrees with the the expected pkcs11 session state.

Cards without a <driver>_check_state function can benefit from this modification,
as sc_lock is now called from sc_pkcs15_check_state. if the lock is not held,
pcsc will get involved and could return card removed, card reset or other error.
sc_lock will also do some SM checks.

These routines try to match the calls as uses for sign, decrypt and derive.

A call to sc_pkcs15_pincache_revalidate could be added to sc_pkcs15_check_state.

This is experimental, and changes may be needed for other cards, or situations.

 Changes to be committed:
	modified:   card-piv.c
	modified:   card.c
	modified:   iso7816.c
	modified:   libopensc.exports
	modified:   opensc.h
	modified:   pkcs15.c
	modified:   ../pkcs11/framework-pkcs15.c
	modified:   ../pkcs11/framework-pkcs15init.c
	modified:   ../pkcs11/pkcs11-session.c
	modified:   ../pkcs11/sc-pkcs11.h