pincache: respect pin_cache_counter
sc_pkcs15_pincache_add was being called at every verify,
which is safe beside reseting usage_counter to 0. That
was causing the cache to be reused above pin_cache_counter.

This doesn't produce any effects by default because
pin_cache_counter is 10 by default and it usually the pin
is not reused 10 times.

Signed-off-by: Nuno Goncalves <nunojpg@gmail.com>